image_extraction/requirements.txt
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
h5py==3.12.1
idna==3.11
ImageIO==2.37.2
lazy_loader==0.4
networkx==3.2.1
numpy==2.0.2
packaging==26.0
pandas==2.2.3
pillow==11.1.0
pyasn1==0.6.2
pycparser==2.23
pyifcb @ git+https://github.com/joefutrelle/pyifcb@4d786eca45c02951bf825fcf35ffecb40d63e867
pysmb==1.2.10
pyspnego==0.12.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rectpack==0.2.2
requests==2.32.3
scikit-image==0.24.0
scipy==1.13.1
six==1.17.0
smbprotocol==1.15.0
tifffile==2024.8.30
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
